Extract parsing to RequestParser
